requirements.txt
pyodbc==5.0.1
pandas==2.0.3
numpy==1.24.3
regex==2023.6.3
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
spacy==3.6.1
streamlit==1.48.0
plotly==5.15.0
sqlalchemy==2.0.19
python-dotenv==1.0.0
faker==19.3.0
cryptography==41.0.7
anthropic>=0.25.0
